+++ helpers/checkHelpers.js
