+++ helpers/controllerWrapper.js
@@ -1,11 +1,13 @@
-const asyncWrapper = (controller) => {
-  return async (req, res, next) => {
-    try {
-      return controller(req, res, next);
-    } catch (error) {
-      next(error);
-    }
-  };
-};
+const asyncWrapper = (callbacks) =>
+  callbacks.map((callback) =>
+    async (req, res, next) => {
+      try {
+        return callback(req, res, next);
+      } catch (error) {
+        next(error);
+      }
+    },
+  )
+;
 
 module.exports = asyncWrapper;
